Move tests for ServerConfigChecks class to separate file
